Enable identifying users after signup
We are using our session GUID to identify anonymous users on
Mixpanel. After signup we are aliasing user with our user ID. This is
not ideal solution since, for some reason, our session GUID is stored in
the permanent cookie so if two users are using same browser that will
not give correct data.