Video tracking data sent to the server
Whenever student finishes watching video or leaves page total number of
seconds is watched to the server. On the server side we updating seconds
watched value and marking element as complete. If user does not leave
the page and replays video new seconds watched value is added to the
existing one.