Sign up user from new landing pages
User is signed up without credit card information. Process assumes that
default subsription plan exist. Default subscription plan is the one
with price 0.0 (free so it does not require credit card info). Procedure
is in other parts unchanged (before and after filters are same for user
and for subscription) except that we do not send subscription data to
Mixpanel since it does not make sence in this case.