Change referred subscription handling logic
With new sign-up process we have to handle customer.subscription.created
event instead of previous update event. We are updating referred signup
only if created subscription is second for this user and previous
subscription is free trial one and is cancelled. Referred subscription
is also bound to the newly created subscription.