Added conditional to subscriptions support in case no yearly plans are present
Removed the links to invioce show page as not sure we can retry authentication on same charge

Updates to invoices and stripe_api_events to ensure invoices are created correctly and correct emails are sent

Updated the stripe handlePayment method to change the subscription back to pending if the 3d_secure auth fails

Added conditional to invoice lists to not render those with status pending. Also returned the link to show invoice to authenticate

Fixed failing specs for stripe_api_events

Removed commented binding.pry

Reworked the layout of the show_invoice page. Moved the ReAuth call to a button and added call to trigger card modal when ReAuth fails

Moved the GA partial to the head and the ifram to top of body. Added conditional to GA partial to ensure the datalayer object is present in development

Added attribute to subscription model to store a guid string used as param on personal_upgrade_complete page to find the correct subscription and then remove the guid to ensure the page cannot be loaded more than once.

Added card details error class to orders stripe form and renamed show_invoice css classes to not clash