(BUG) Coupon validate in change plan.
We had some errors coming from the validate coupon in change plan.

This was occurred when a user when to change plan but doesn't see any other plan available to change.

I add a check to not show the option to user if no plan is available to change.
Also add a check in javascript to not do any ajax request if we're not passing planId as param.